+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed multi-asset output handling in `debug-pilet` with `piral-cli-webpack` (#258)
 * Added basic support for mounting modules in `piral-ng`
 * Added ability to add new menu types in `piral-menu`
+* Added options for configuration of the `piletApi` in the `.piralrc`
 
 ## 0.11.7 (June 22, 2020)